feat: add test execution duration
